Reduced the size of `with open(..., 'w')` blocks
`with open(..., 'w')` sets the opened file to be 0 bytes. This is useful
when overwriting data, but means that data will be lost if we're not
careful and an exception occurs. This has happened several times during
development.

We've now reduced these blocks to only the parts where a file is
actually written, to prevent data from being lost. This does increase
the likelyhood of a race condition between opens, so we'll need to look
info file locking.